Fixed warning in fcn_18 ref unable to unlink directory - Fixed more linting issues

+++ src/functions/fcn_18_unlinkArchiveOld.php
@@ -14,8 +14,10 @@ function fcn_18_unlinkArchiveOld($path, $logger) // $strBackupFolder
             $filelastmodified = filemtime($path . "/" . $file);
             //3 days * 24 hours in a day * 3600 seconds per hour
             if ((time() - $filelastmodified) > 1 * 3600) {
+                if ($file != "." && $file != "..") {
                 unlink($path . "/" . $file);
                 $logger->info("File " . $file . " removed from " . $path . "");
+                }
             }
         }
         closedir($handle);
